refactor: <un-button/> standards alignment
Refactored to be in the light dom, so that native browser behaviours work more consistently.

Removed the `text` attribute: labels are just set as the child text content.

Move "type" to "variant" so that the "type" attribute can be equivalent to <button type /> which has "button", "submit" and "reset" as valid values.

refactor: default <button /> to type=button

otherwise it defaults to type=submit, which causes strange behaviour

fix: button dom management

this was breaking lit rendering further up the chain